# apis/api/swagger/swagger.yaml
components:
...
    schemas:
        System:
            properties:
                _id:
                    type: string
                referencedParts:
                    items:
                        type: string
                    type: array
                description:
                    type: string
                    description: 'A full description of the system'
                summary:
                    type: string
                    description: 'Short summary of the system'
                name:
                    type: string
                    description: 'The name of the system'
            required:
                - _id
                - referencedParts
                - description
                - summary
                - name
            type: object
...
        Service:
            properties:
                _id:
                    type: string
                newParts:
                    items:
                        type: string
                    type: array
                partsServiced:
                    items:
                        type: string
                    type: array
                    description: 'parts being affected by the service'
                systems:
                    items:
                        type: string
                    type: array
                    description: 'Systems the Service is affecting'
                steps:
                    items:
                        type: string
                    type: array
                description:
                    type: string
                    description: 'Description of the service'
                name:
                    type: string
                    description: 'The name of the Service'
            required:
                - _id
                - newParts
                - partsServiced
                - systems
                - steps
                - description
                - name
            type: object
        Pick_Service.Exclude_keyofService._id:
            properties:
                name:
                    type: string
                    description: 'The name of the Service'
                description:
                    type: string
                    description: 'Description of the service'
                steps:
                    items:
                        type: string
                    type: array
                systems:
                    items:
                        type: string
                    type: array
                    description: 'Systems the Service is affecting'
                partsServiced:
                    items:
                        type: string
                    type: array
                    description: 'parts being affected by the service'
                newParts:
                    items:
                        type: string
                    type: array
            required:
                - name
                - description
                - steps
                - systems
                - partsServiced
                - newParts
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        OmitOverride_Service._id_:
            $ref: '#/components/schemas/Pick_Service.Exclude_keyofService._id'
        ServiceCreate:
            $ref: '#/components/schemas/OmitOverride_Service._id_'
        Pick_System.Exclude_keyofSystem._id:
            properties:
                name:
                    type: string
                    description: 'The name of the system'
                description:
                    type: string
                    description: 'A full description of the system'
                summary:
                    type: string
                    description: 'Short summary of the system'
                referencedParts:
                    items:
                        type: string
                    type: array
            required:
                - name
                - description
                - summary
                - referencedParts
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        OmitOverride_System._id_:
            $ref: '#/components/schemas/Pick_System.Exclude_keyofSystem._id'
        SystemCreate:
            $ref: '#/components/schemas/OmitOverride_System._id_'
...
    securitySchemes:
        jwt:
            type: apiKey
            name: authorization
            in: header
info:
    title: '@helium/ts-tsoa-api'
    version: 0.0.2
    description: 'Helium API with TSOA generator and openAPI sdk generator'
    contact: {}
openapi: 3.0.0
paths:
...
    /service/find/all:
        post:
            operationId: AdminFindservice
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Service'
                                type: array
            tags:
                - service
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'the page that is returned'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 1
                        format: int32
                        type: integer
                -
                    description: 'the number of records in the page'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 10
                        format: int32
                        type: integer
    /service/find:
        post:
            operationId: Findservices
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Service'
                                type: array
            tags:
                - service
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'page needs description'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 0
                        format: int32
                        type: integer
                -
                    description: 'count needs description'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 20
                        format: int32
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                conditions:
                                    additionalProperties: true
                            required:
                                - conditions
                            type: object
    '/service/{serviceId}':
        get:
            operationId: Getservice
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Service'
            tags:
                - service
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: serviceId
                    required: true
                    schema:
                        type: string
        put:
            operationId: Updateservice
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Service'
            tags:
                - service
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: serviceId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ServiceCreate'
        delete:
            operationId: Deleteservice
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: boolean
            tags:
                - service
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: serviceId
                    required: true
                    schema:
                        type: string
    /service:
        post:
            operationId: Createservice
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Service'
            tags:
                - service
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ServiceCreate'
    /system/find/all:
        post:
            operationId: AdminFindsystem
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/System'
                                type: array
            tags:
                - system
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'the page that is returned'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 1
                        format: int32
                        type: integer
                -
                    description: 'the number of records in the page'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 10
                        format: int32
                        type: integer
    /system/find:
        post:
            operationId: Findsystems
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/System'
                                type: array
            tags:
                - system
            security:
                -
                    jwt:
                        - admin
            parameters:
                -
                    description: 'page needs description'
                    in: header
                    name: page
                    required: false
                    schema:
                        default: 0
                        format: int32
                        type: integer
                -
                    description: 'count needs description'
                    in: header
                    name: count
                    required: false
                    schema:
                        default: 20
                        format: int32
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            properties:
                                conditions:
                                    additionalProperties: true
                            required:
                                - conditions
                            type: object
    '/system/{systemId}':
        get:
            operationId: Getsystem
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/System'
            tags:
                - system
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: systemId
                    required: true
                    schema:
                        type: string
        put:
            operationId: Updatesystem
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/System'
            tags:
                - system
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: systemId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SystemCreate'
        delete:
            operationId: Deletesystem
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: boolean
            tags:
                - system
            security:
                -
                    jwt:
                        - owner
            parameters:
                -
                    in: path
                    name: systemId
                    required: true
                    schema:
                        type: string
    /system:
        post:
            operationId: Createsystem
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/System'
            tags:
                - system
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SystemCreate'
...
servers:
    -
        url: 'http://localhost:3002/v1'